refactor: move inference metadata to seperate classes and modules
