+++ metadata/src/anemoi/metadata/versions/v1.py
@@ -15,12 +15,6 @@ holds a strictly-typed :class:`InferenceMetadata` block alongside permissive
 ``dict`` sections for training, dataset, environment, and provenance data that
 do not need to be validated at this layer.
 
-The nested per-dataset models (:class:`DatasetInferenceConfig`,
-:class:`DataIndices`, :class:`VariableTypes`, :class:`TimestepConfig`,
-:class:`TensorShapes`) preserve unknown fields for forward compatibility. This
-allows newer checkpoint writers to add fields without breaking older readers.
-Strict validation will be enforced at write time from V2 onwards.
-
 V1 only handles checkpoints that already have a ``metadata_inference`` block.
 Legacy checkpoints (no ``metadata_inference``, no ``schema_version``) are
 handled by :class:`~anemoi.metadata.versions.v0.MetadataV0`.  A migration
@@ -31,241 +25,17 @@ from V0 to V1 is registered in
 from datetime import datetime
 from typing import Any
 
-from pydantic import BaseModel
 from pydantic import ConfigDict
 from pydantic import Field
-from pydantic import model_validator
 
 from ..base import MetadataContract
 from ..registry import MetadataRegistry
-
-
-class TimestepConfig(BaseModel):
-    """Temporal stepping configuration for a dataset.
-
-    Captures the frequency string and the relative date index arrays that
-    describe which input/output time-steps the model was trained with.
-
-    Unknown fields are preserved for forward compatibility; newer checkpoint
-    writers may add fields that older readers do not recognise. Strict
-    validation will be enforced at write time from V2 onwards.
-
-    Attributes
-    ----------
-    timestep : str
-        Frequency string, e.g. ``"6h"``.
-    input_relative_date_indices : list[int]
-        Relative date indices used as model inputs (e.g. ``[-1, 0]``).
-    output_relative_date_indices : list[int]
-        Relative date indices produced as model outputs (e.g. ``[1]``).
-    relative_date_indices_training : list[int]
-        Full set of relative date indices seen during training.
-    """
-
-    model_config = ConfigDict(extra="allow", frozen=True)
-
-    timestep: str
-    input_relative_date_indices: list[int]
-    output_relative_date_indices: list[int]
-    relative_date_indices_training: list[int]
-
-
-class DataIndices(BaseModel):
-    """Mapping from variable names to tensor indices.
-
-    Unknown fields are preserved for forward compatibility; newer checkpoint
-    writers may add fields that older readers do not recognise. Strict
-    validation will be enforced at write time from V2 onwards.
-
-    Attributes
-    ----------
-    input : dict[str, int]
-        Mapping of variable name to its index in the input tensor.
-    output : dict[str, int]
-        Mapping of variable name to its index in the output tensor.
-    """
-
-    model_config = ConfigDict(extra="allow", frozen=True)
-
-    input: dict[str, int]
-    output: dict[str, int]
-
-
-class VariableTypes(BaseModel):
-    """Categorisation of variables by their role in the model.
-
-    Unknown fields are preserved for forward compatibility; newer checkpoint
-    writers may add fields that older readers do not recognise. Strict
-    validation will be enforced at write time from V2 onwards.
-
-    Attributes
-    ----------
-    forcing : list[str]
-        Variables that are provided as external forcings (not predicted).
-    target : list[str]
-        Variables that the model is trained to predict.
-    prognostic : list[str]
-        Variables that are both input and output (stepped forward in time).
-    diagnostic : list[str]
-        Variables that are output-only diagnostics (not fed back as input).
-    """
-
-    model_config = ConfigDict(extra="allow", frozen=True)
-
-    forcing: list[str] = Field(default_factory=list)
-    target: list[str] = Field(default_factory=list)
-    prognostic: list[str] = Field(default_factory=list)
-    diagnostic: list[str] = Field(default_factory=list)
-
-
-class TensorShapes(BaseModel):
-    """Shape metadata for the model's input/output tensors.
-
-    Unknown fields are preserved for forward compatibility; newer checkpoint
-    writers may add fields that older readers do not recognise. Strict
-    validation will be enforced at write time from V2 onwards.
-
-    Attributes
-    ----------
-    variables : int
-        Number of variables (channels) in the tensor.
-    input_timesteps : int
-        Number of input time-steps stacked along the time dimension.
-    ensemble : int
-        Ensemble size; defaults to ``1`` for deterministic models.
-    grid : int or None
-        Number of grid points, or ``None`` when not applicable.
-    """
-
-    model_config = ConfigDict(extra="allow", frozen=True)
-
-    variables: int
-    input_timesteps: int
-    ensemble: int = 1
-    grid: int | None = None
-
-
-class DatasetInferenceConfig(BaseModel):
-    """Inference configuration for a single named dataset.
-
-    Bundles together the index mappings, variable categorisation, temporal
-    stepping, and tensor-shape information needed by inference for one dataset.
-
-    Unknown fields are preserved for forward compatibility; newer checkpoint
-    writers may add fields that older readers do not recognise. Strict
-    validation will be enforced at write time from V2 onwards.
-
-    Attributes
-    ----------
-    data_indices : DataIndices
-        Variable-name-to-tensor-index mappings for input and output.
-    variable_types : VariableTypes
-        Categorisation of variables by role.
-    timesteps : TimestepConfig
-        Temporal stepping configuration.
-    shapes : TensorShapes
-        Tensor shape metadata.
-    """
-
-    model_config = ConfigDict(extra="allow", frozen=True)
-
-    data_indices: DataIndices
-    variable_types: VariableTypes
-    timesteps: TimestepConfig
-    shapes: TensorShapes
-
-
-class InferenceMetadata(BaseModel):
-    """Top-level inference metadata written by training.
-
-    This model is the authoritative source of truth consumed by inference at
-    runtime.  It supports two input shapes:
-
-    1. **Structured** - a dict that already contains a ``"datasets"`` key
-       mapping dataset names to their per-dataset configs.
-    2. **Flat** - a dict where scalar fields (``seed``, ``run_id``, ``task``,
-       ``dataset_names``) sit alongside per-dataset sub-dicts keyed by the
-       names listed in ``dataset_names``.  The ``@model_validator`` reshapes
-       this into the structured form before validation.
-
-    Unknown extra keys at the root level are preserved (``extra="allow"``) so
-    that future fields added by training do not break existing inference code.
-
-    Attributes
-    ----------
-    seed : int
-        Random seed used during training.
-    run_id : str
-        Unique identifier for the training run.
-    task : str or None
-        Optional task label (e.g. ``"forecaster"``).
-    dataset_names : list[str]
-        Ordered list of dataset names referenced by this checkpoint.
-    datasets : dict[str, DatasetInferenceConfig]
-        Per-dataset inference configuration, keyed by dataset name.
-    """
-
-    model_config = ConfigDict(extra="allow", frozen=True)
-
-    seed: int
-    run_id: str
-    task: str | None = None
-    dataset_names: list[str]
-    datasets: dict[str, DatasetInferenceConfig]
-
-    @model_validator(mode="before")
-    @classmethod
-    def _restructure_flat_checkpoint(cls, values: Any) -> Any:
-        """Reshape a flat checkpoint dict into the structured form.
-
-        If the incoming data already contains a ``"datasets"`` key the dict is
-        returned unchanged.  Otherwise the validator extracts the known scalar
-        fields and collects the remaining keys that are listed in
-        ``dataset_names`` into a ``"datasets"`` sub-dict.
-
-        Parameters
-        ----------
-        values : Any
-            Raw input passed to the model constructor.
-
-        Returns
-        -------
-        Any
-            Either the original value (if not a dict or already structured) or
-            a restructured dict with a ``"datasets"`` key.
-        """
-        if not isinstance(values, dict):
-            return values
-
-        # Already in structured form - nothing to do.
-        if "datasets" in values:
-            return values
-
-        _SCALAR_KEYS = {"seed", "run_id", "task", "dataset_names"}
-
-        dataset_names: list[str] = values.get("dataset_names", [])
-
-        scalars: dict[str, Any] = {k: v for k, v in values.items() if k in _SCALAR_KEYS}
-
-        missing = [k for k in dataset_names if k not in values]
-        if missing:
-            msg = (
-                f"metadata_inference references datasets {missing} in "
-                f"'dataset_names' but no corresponding entries exist in the metadata."
-            )
-            raise ValueError(msg)
-
-        datasets: dict[str, Any] = {k: values[k] for k in dataset_names}
-
-        # Preserve any extra keys that are neither scalars nor dataset entries
-        # so that extra="allow" can capture them at the root level.
-        extras: dict[str, Any] = {k: v for k, v in values.items() if k not in _SCALAR_KEYS and k not in dataset_names}
-
-        return {**scalars, "datasets": datasets, **extras}
+from .inference import InferenceMetadata
+from .inference.contract_fullfillment import InferenceMetadataFullfillment
 
 
 @MetadataRegistry.register("1.0")
-class MetadataV1(MetadataContract):
+class MetadataV1(InferenceMetadataFullfillment, MetadataContract):
     """Version 1.0 metadata schema.
 
     The top-level container for all checkpoint metadata.  The
@@ -312,178 +82,6 @@ class MetadataV1(MetadataContract):
     environment: dict[str, Any] = Field(default_factory=dict)
     provenance: dict[str, Any] = Field(default_factory=dict)
 
-    # ------------------------------------------------------------------
-    # Internal helpers
-    # ------------------------------------------------------------------
-
-    def _resolve_dataset(self, dataset_name: str | None) -> DatasetInferenceConfig:
-        """Return the :class:`DatasetInferenceConfig` for a named dataset.
-
-        Parameters
-        ----------
-        dataset_name : str | None
-            Dataset name to look up.  When ``None`` the first entry in
-            ``metadata_inference.dataset_names`` is used.
-
-        Returns
-        -------
-        DatasetInferenceConfig
-            Per-dataset inference configuration.
-
-        Raises
-        ------
-        KeyError
-            If *dataset_name* is not present in ``metadata_inference.datasets``.
-        IndexError
-            If ``dataset_names`` is empty and *dataset_name* is ``None``.
-        """
-        if dataset_name is None:
-            dataset_name = self.metadata_inference.dataset_names[0]
-        return self.metadata_inference.datasets[dataset_name]
-
-    # ------------------------------------------------------------------
-    # MetadataContract contract implementation
-    # ------------------------------------------------------------------
-
-    def get_dataset_names(self) -> list[str]:
-        """Return the ordered list of dataset names.
-
-        Returns
-        -------
-        list[str]
-            Dataset names as recorded in ``metadata_inference``.
-        """
-        return self.metadata_inference.dataset_names
-
-    def get_task(self) -> str | None:
-        """Return the task label, or ``None`` if not set.
-
-        Returns
-        -------
-        str or None
-            Task label (e.g. ``"forecaster"``).
-        """
-        return self.metadata_inference.task
-
-    def get_timestep(self, dataset_name: str | None = None) -> str:
-        """Return the timestep frequency string for a dataset.
-
-        Parameters
-        ----------
-        dataset_name : str | None, optional
-            Dataset to query.  Defaults to the first dataset.
-
-        Returns
-        -------
-        str
-            Frequency string (e.g. ``"6h"``).
-        """
-        return self._resolve_dataset(dataset_name).timesteps.timestep
-
-    def get_input_relative_date_indices(self, dataset_name: str | None = None) -> list[int]:
-        """Return input relative date indices for a dataset.
-
-        Parameters
-        ----------
-        dataset_name : str | None, optional
-            Dataset to query.  Defaults to the first dataset.
-
-        Returns
-        -------
-        list[int]
-            Relative date indices used as model inputs (e.g. ``[-1, 0]``).
-        """
-        return self._resolve_dataset(dataset_name).timesteps.input_relative_date_indices
-
-    def get_output_relative_date_indices(self, dataset_name: str | None = None) -> list[int]:
-        """Return output relative date indices for a dataset.
-
-        Parameters
-        ----------
-        dataset_name : str | None, optional
-            Dataset to query.  Defaults to the first dataset.
-
-        Returns
-        -------
-        list[int]
-            Relative date indices produced as model outputs (e.g. ``[1]``).
-        """
-        return self._resolve_dataset(dataset_name).timesteps.output_relative_date_indices
-
-    def get_variable_indices(self, dataset_name: str | None = None) -> dict[str, int]:
-        """Return input variable name to tensor index mapping.
-
-        Parameters
-        ----------
-        dataset_name : str | None, optional
-            Dataset to query.  Defaults to the first dataset.
-
-        Returns
-        -------
-        dict[str, int]
-            Mapping of variable name to its index in the input tensor.
-        """
-        return self._resolve_dataset(dataset_name).data_indices.input
-
-    def get_output_variable_indices(self, dataset_name: str | None = None) -> dict[str, int]:
-        """Return output variable name to tensor index mapping.
-
-        Parameters
-        ----------
-        dataset_name : str | None, optional
-            Dataset to query.  Defaults to the first dataset.
-
-        Returns
-        -------
-        dict[str, int]
-            Mapping of variable name to its index in the output tensor.
-        """
-        return self._resolve_dataset(dataset_name).data_indices.output
-
-    def get_variable_types(self, dataset_name: str | None = None) -> dict[str, list[str]]:
-        """Return variable categories by role.
-
-        Parameters
-        ----------
-        dataset_name : str | None, optional
-            Dataset to query.  Defaults to the first dataset.
-
-        Returns
-        -------
-        dict[str, list[str]]
-            Dictionary mapping category names (``"forcing"``, ``"prognostic"``,
-            ``"diagnostic"``, ``"target"``) to lists of variable names.
-        """
-        vt = self._resolve_dataset(dataset_name).variable_types
-        return {
-            "forcing": list(vt.forcing),
-            "prognostic": list(vt.prognostic),
-            "diagnostic": list(vt.diagnostic),
-            "target": list(vt.target),
-        }
-
-    def get_tensor_shapes(self, dataset_name: str | None = None) -> dict[str, Any]:
-        """Return tensor shape metadata as a plain dict.
-
-        Parameters
-        ----------
-        dataset_name : str | None, optional
-            Dataset to query.  Defaults to the first dataset.
-
-        Returns
-        -------
-        dict[str, Any]
-            Shape metadata with keys ``"variables"``, ``"input_timesteps"``,
-            ``"ensemble"``, and ``"grid"``.
-        """
-        shapes = self._resolve_dataset(dataset_name).shapes
-        return {
-            "variables": shapes.variables,
-            "input_timesteps": shapes.input_timesteps,
-            "ensemble": shapes.ensemble,
-            "grid": shapes.grid,
-        }
-
     def get_variables_metadata(self, dataset_name: str | None = None) -> dict[str, dict[str, Any]]:
         """Return per-variable metadata from the permissive dataset section.
 
@@ -518,23 +116,6 @@ class MetadataV1(MetadataContract):
                 result[name] = {**result[name], "constant_in_time": True}
 
         return result
-
-    def get_grid_points(self, dataset_name: str | None = None) -> int | None:
-        """Return the number of grid points from the typed shapes block.
-
-        In V1, read from ``shapes.grid`` of the resolved dataset.
-
-        Parameters
-        ----------
-        dataset_name : str | None, optional
-            Dataset to query.  Defaults to the first dataset.
-
-        Returns
-        -------
-        int or None
-            Number of grid points, or ``None`` if not recorded.
-        """
-        return self._resolve_dataset(dataset_name).shapes.grid
 
     def get_data_request(self, dataset_name: str | None = None) -> dict[str, Any]:
         """Return data request parameters from the permissive dataset section.
